+++ classes/event/assignment_restored.php
@@ -70,10 +70,10 @@ final class assignment_restored extends \core\event\base {
     /**
      * Get URL related to the action.
      *
-     * @return \moodle_url
+     * @return \core\url
      */
     public function get_url() {
-        return new \moodle_url('/admin/tool/mucertify/management/assignment.php', ['id' => $this->objectid]);
+        return new \core\url('/admin/tool/mucertify/management/assignment.php', ['id' => $this->objectid]);
     }
 
     /**