+++ classes/reportbuilder/local/systemreports/assignments.php
@@ -27,7 +27,6 @@ use core_reportbuilder\system_report;
 use core_reportbuilder\local\helpers\database;
 use core_reportbuilder\local\helpers\user_profile_fields;
 use lang_string;
-use moodle_url;
 
 /**
  * Embedded certification assignments report.
@@ -108,7 +107,7 @@ final class assignments extends system_report {
         $column
             ->add_fields("$assignmentalias.id")
             ->add_callback(static function (string $fullname, \stdClass $row): string {
-                $url = new \moodle_url('/admin/tool/mucertify/management/assignment.php', ['id' => $row->id]);
+                $url = new \core\url('/admin/tool/mucertify/management/assignment.php', ['id' => $row->id]);
                 return \html_writer::link($url, $fullname);
             });
         $this->add_column($column);
@@ -181,7 +180,7 @@ final class assignments extends system_report {
 
         $certification = $this->certification;
 
-        $url = new moodle_url('/admin/tool/mucertify/management/assignment_update.php', ['id' => ':id']);
+        $url = new \core\url('/admin/tool/mucertify/management/assignment_update.php', ['id' => ':id']);
         $link = new \tool_mulib\output\ajax_form\link($url, get_string('assignment_update', 'tool_mucertify'), 'i/settings');
         $this->add_action($link->create_report_action()
             ->add_callback(static function (\stdclass $row) use ($certification): bool {
@@ -210,7 +209,7 @@ final class assignments extends system_report {
                 return $sourceclass::is_assignment_update_possible($certification, $source, $assignment);
             }));
 
-        $url = new moodle_url('/admin/tool/mucertify/management/assignment_archive.php', ['id' => ':id']);
+        $url = new \core\url('/admin/tool/mucertify/management/assignment_archive.php', ['id' => ':id']);
         $link = new \tool_mulib\output\ajax_form\link($url, get_string('assignment_archive', 'tool_mucertify'), 'i/lock');
         $this->add_action($link->create_report_action()
             ->add_callback(static function (\stdclass $row) use ($certification): bool {
@@ -239,7 +238,7 @@ final class assignments extends system_report {
                 return $sourceclass::is_assignment_archive_possible($certification, $source, $assignment);
             }));
 
-        $url = new moodle_url('/admin/tool/mucertify/management/assignment_restore.php', ['id' => ':id']);
+        $url = new \core\url('/admin/tool/mucertify/management/assignment_restore.php', ['id' => ':id']);
         $link = new \tool_mulib\output\ajax_form\link($url, get_string('assignment_restore', 'tool_mucertify'), 'i/unlock');
         $this->add_action($link->create_report_action()
             ->add_callback(static function (\stdclass $row) use ($certification): bool {
@@ -268,7 +267,7 @@ final class assignments extends system_report {
                 return $sourceclass::is_assignment_restore_possible($certification, $source, $assignment);
             }));
 
-        $url = new moodle_url('/admin/tool/mucertify/management/assignment_delete.php', ['id' => ':id']);
+        $url = new \core\url('/admin/tool/mucertify/management/assignment_delete.php', ['id' => ':id']);
         $link = new \tool_mulib\output\ajax_form\link($url, get_string('assignment_delete', 'tool_mucertify'), 'i/delete');
         $this->add_action($link->create_report_action(['class' => 'text-danger'])
             ->add_callback(static function (\stdclass $row) use ($certification): bool {