+++ management/certification_settings_edit1.php
@@ -47,13 +47,13 @@ $certification = $DB->get_record('tool_mucertify_certification', ['id' => $id], 
 $context = context::instance_by_id($certification->contextid);
 require_capability('tool/mucertify:edit', $context);
 
-$currenturl = new moodle_url('/admin/tool/mucertify/management/certification_settings_edit1.php', ['id' => $id]);
+$currenturl = new \core\url('/admin/tool/mucertify/management/certification_settings_edit1.php', ['id' => $id]);
 $PAGE->set_context($context);
 $PAGE->set_url($currenturl);
 
 $form = new \tool_mucertify\local\form\certification_settings_edit1(null, ['certification' => $certification, 'context' => $context]);
 
-$returnurl = new moodle_url('/admin/tool/mucertify/management/certification_settings.php', ['id' => $certification->id]);
+$returnurl = new \core\url('/admin/tool/mucertify/management/certification_settings.php', ['id' => $certification->id]);
 
 if ($form->is_cancelled()) {
     $form->ajax_form_cancelled($returnurl);