+++ tests/phpunit/event/certification_updated_test.php
@@ -70,7 +70,7 @@ final class certification_updated_test extends \advanced_testcase {
         $this->assertSame('tool_mucertify_certification', $event->objecttable);
         $this->assertSame('Certification updated', $event::get_name());
         $description = $event->get_description();
-        $certificationurl = new \moodle_url('/admin/tool/mucertify/management/certification.php', ['id' => $certification->id]);
+        $certificationurl = new \core\url('/admin/tool/mucertify/management/certification.php', ['id' => $certification->id]);
         $this->assertSame($certificationurl->out(false), $event->get_url()->out(false));
 
         $sink = $this->redirectEvents();
@@ -92,7 +92,7 @@ final class certification_updated_test extends \advanced_testcase {
         $this->assertSame('tool_mucertify_certification', $event->objecttable);
         $this->assertSame('Certification updated', $event::get_name());
         $description = $event->get_description();
-        $certificationurl = new \moodle_url('/admin/tool/mucertify/management/certification.php', ['id' => $certification->id]);
+        $certificationurl = new \core\url('/admin/tool/mucertify/management/certification.php', ['id' => $certification->id]);
         $this->assertSame($certificationurl->out(false), $event->get_url()->out(false));
 
         $sink = $this->redirectEvents();
@@ -114,7 +114,7 @@ final class certification_updated_test extends \advanced_testcase {
         $this->assertSame('tool_mucertify_certification', $event->objecttable);
         $this->assertSame('Certification updated', $event::get_name());
         $description = $event->get_description();
-        $certificationurl = new \moodle_url('/admin/tool/mucertify/management/certification.php', ['id' => $certification->id]);
+        $certificationurl = new \core\url('/admin/tool/mucertify/management/certification.php', ['id' => $certification->id]);
         $this->assertSame($certificationurl->out(false), $event->get_url()->out(false));
     }
 }