+++ tests/phpunit/local/management_test.php
@@ -68,7 +68,7 @@ final class management_test extends \advanced_testcase {
         $this->assertNull(management::get_management_url());
 
         $this->setUser($admin);
-        $expected = new \moodle_url('/admin/tool/mucertify/management/index.php');
+        $expected = new \core\url('/admin/tool/mucertify/management/index.php');
         $this->assertSame((string)$expected, (string)management::get_management_url());
 
         $this->setUser($manager);
@@ -101,11 +101,11 @@ final class management_test extends \advanced_testcase {
         role_assign($viewerroleid, $viewer1->id, $tenantcatcontext->id);
 
         $this->setUser($viewer0);
-        $expected = new \moodle_url('/admin/tool/mucertify/management/index.php');
+        $expected = new \core\url('/admin/tool/mucertify/management/index.php');
         $this->assertSame((string)$expected, (string)management::get_management_url());
 
         $this->setUser($viewer1);
-        $expected = new \moodle_url('/admin/tool/mucertify/management/index.php', ['contextid' => $tenantcatcontext->id]);
+        $expected = new \core\url('/admin/tool/mucertify/management/index.php', ['contextid' => $tenantcatcontext->id]);
         $this->assertSame((string)$expected, (string)management::get_management_url());
     }
 
@@ -156,14 +156,14 @@ final class management_test extends \advanced_testcase {
 
         $PAGE = new \moodle_page();
         management::setup_index_page(
-            new \moodle_url('/admin/tool/mucertify/management/index.php'),
+            new \core\url('/admin/tool/mucertify/management/index.php'),
             $syscontext
         );
 
         $this->setUser($user);
         $PAGE = new \moodle_page();
         management::setup_index_page(
-            new \moodle_url('/admin/tool/mucertify/management/index.php'),
+            new \core\url('/admin/tool/mucertify/management/index.php'),
             $syscontext
         );
     }
@@ -181,7 +181,7 @@ final class management_test extends \advanced_testcase {
 
         $PAGE = new \moodle_page();
         management::setup_certification_page(
-            new \moodle_url('/admin/tool/mucertify/management/new.php'),
+            new \core\url('/admin/tool/mucertify/management/new.php'),
             $syscontext,
             $certification1,
             'certification_general'
@@ -190,7 +190,7 @@ final class management_test extends \advanced_testcase {
         $this->setUser($user);
         $PAGE = new \moodle_page();
         management::setup_certification_page(
-            new \moodle_url('/admin/tool/mucertify/management/new.php'),
+            new \core\url('/admin/tool/mucertify/management/new.php'),
             $syscontext,
             $certification1,
             'certification_general'